Table overhaul
