+++ cfme/cloud/stack.py
@@ -1,6 +1,7 @@
 import ui_navigate as nav
 import cfme.fixtures.pytest_selenium as sel
-from cfme.web_ui import Quadicon, Region, SplitTable, flash, Form, fill, form_buttons, Table
+from cfme.web_ui import Quadicon, Region, flash, Form, fill, form_buttons
+from cfme.web_ui.tables import Table, Split
 from utils.pretty import Pretty
 from functools import partial
 from cfme.web_ui import toolbar as tb
@@ -15,12 +16,12 @@ details_page = Region(infoblock_type='detail')
 cfg_btn = partial(tb.select, "Configuration")
 pol_btn = partial(tb.select, 'Policy')
 lifecycle_btn = partial(tb.select, 'Lifecycle')
-output_table = lambda: version.pick(
-    {'5.5': Table('//div[@id="list_grid"]/table'),
-    '5.4': SplitTable(
-        ('//*[@id="list_grid"]//table[contains(@class, "hdr")]/tbody', 1),
-        ('//*[@id="list_grid"]//table[contains(@class, "obj")]/tbody', 1))}
-)
+output_table = Table.create({
+    version.LOWEST: Split(
+        '//*[@id="list_grid"]//table[contains(@class, "hdr")]/tbody',
+        '//*[@id="list_grid"]//table[contains(@class, "obj")]/tbody',
+        1, 1),
+    '5.5': '//div[@id="list_grid"]/table'})
 
 edit_tags_form = Form(
     fields=[
@@ -81,7 +82,7 @@ class Stack(Pretty):
         sel.force_navigate('clouds_stack', context={'stack': self})
         sel.click(details_page.infoblock.element("Relationships", "Outputs"))
         cells = {'Key': "WebsiteURL"}
-        output_table().click_rows_by_cells(cells, "Key", True)
+        output_table.click_rows_by_cells(cells, "Key", True)
 
     def nav_to_resources_link(self):
         sel.force_navigate('clouds_stack', context={'stack': self})
