+++ cfme/web_ui/__init__.py
@@ -184,863 +184,6 @@ def get_context_current_page():
     return stripped[stripped.find('/'):stripped.rfind('?')]
 
 
-def _convert_header(header):
-    """Convers header cell text into something usable as an identifier.
-
-    Static method which replaces spaces in headers with underscores and strips out
-    all other characters to give an identifier.
-
-    Args:
-        header: A header name to be converted.
-
-    Returns: A string holding the converted header.
-    """
-    return re.sub('[^0-9a-zA-Z_]+', '', header.replace(' ', '_')).lower()
-
-
-class CachedTableHeaders(object):
-    """the internal cache of headers
-
-    This allows columns to be moved and the Table updated. The :py:attr:`headers` stores
-    the header cache element and the list of headers are stored in _headers. The
-    attribute header_indexes is then created, before finally creating the items
-    attribute.
-    """
-    def __init__(self, table):
-        self.headers = sel.elements('td | th', root=table.header_row)
-        self.indexes = {
-            _convert_header(cell.text): index
-            for index, cell in enumerate(self.headers)}
-
-
-class Table(Pretty):
-    """
-    Helper class for Table/List objects
-
-    Turns CFME custom Table/Lists into iterable objects using a generator.
-
-    Args:
-        table_locator: locator pointing to a table element with child thead and tbody elements
-            representing that table's header and body row containers
-        header_offset: In the case of a padding table row above the header, the row offset
-            can be used to skip rows in ``<thead>`` to locate the correct header row. This offset
-            is 1-indexed, not 0-indexed, so an offset of 1 is the first child row element
-        body_offset: In the case of a padding table row above the body rows, the row offset
-            can be used to skip rows in ``<ttbody>`` to locate the correct header row. This offset
-            is 1-indexed, not 0-indexed, so an offset of 1 is the first child row element
-
-    Attributes:
-        header_indexes: A dict of header names related to their int index as a column.
-
-    Usage:
-
-        table = Table('//div[@id="prov_pxe_img_div"]//table')
-
-    The HTML code for the table looks something like this::
-
-        <div id="prov_pxe_img_div">
-          <table>
-              <thead>
-                  <tr>
-                      <th>Name</th>
-                      <th>Animal</th>
-                      <th>Size</th>
-                  </tr>
-              </thead>
-              <tbody>
-                  <tr>
-                      <td>John</td>
-                      <td>Monkey</td>
-                      <td>Small</td>
-                  </tr>
-                  <tr>
-                      <td>Mike</td>
-                      <td>Tiger</td>
-                      <td>Large</td>
-                  </tr>
-              </tbody>
-          </table>
-        </div>
-
-    We can now click on an element in the list like so, by providing the column
-    name and the value that we are searching for::
-
-        table.click_cell('name', 'Mike')
-
-    We can also perform the same, by using the index of the column, like so::
-
-        table.click_cell(1, 'Tiger')
-
-    Additionally, the rows of a table can be iterated over, and that row's columns can be accessed
-    by name or index (left to right, 0-index)::
-
-        for row in table.rows()
-            # Get the first cell in the row
-            row[0]
-            # Get the row's contents for the column with header 'Row Name'
-            # All of these will work, though the first is preferred
-            row.row_name, row['row_name'], row['Row Name']
-
-    When doing bulk opererations, such as selecting rows in a table based on their content,
-    the ``*_by_cells`` methods are able to find matching row much more quickly than iterating,
-    as the work can be done with fewer selenium calls.
-
-        * :py:meth:`find_rows_by_cells`
-        * :py:meth:`find_row_by_cells`
-        * :py:meth:`click_rows_by_cells`
-        * :py:meth:`click_row_by_cells`
-
-    Note:
-
-        A table is defined by the containers of the header and data areas, and offsets to them.
-        This allows a table to include one or more padding rows above the header row. In
-        the example above, there is no padding row, as our offset values are set to 0.
-
-    """
-
-    pretty_attrs = ['_loc']
-
-    def __init__(self, table_locator, header_offset=0, body_offset=0):
-        self._headers = None
-        self._header_indexes = None
-        self._loc = table_locator
-        self.header_offset = int(header_offset)
-        self.body_offset = int(body_offset)
-
-    @property
-    def header_row(self):
-        """Property representing the ``<tr>`` element that contains header cells"""
-        # thead/tr containing header data
-        # xpath is 1-indexed, so we need to add 1 to the offset to get the correct row
-        return sel.element('thead/tr[%d]' % (self.header_offset + 1), root=sel.element(self))
-
-    @property
-    def body(self):
-        """Property representing the ``<tbody>`` element that contains body rows"""
-        # tbody containing body rows
-        return sel.element('tbody', root=sel.element(self))
-
-    @cached_property
-    def _headers_cache(self):
-        return CachedTableHeaders(self)
-
-    def _update_cache(self):
-        """refresh the cache in case we know its stale"""
-        try:
-            del self._headers_cache
-        except AttributeError:
-            pass  # it's not cached, dont try to be eager
-        else:
-            self._headers_cache
-
-    @property
-    def headers(self):
-        """List of ``<td>`` or ``<th>`` elements in :py:attr:`header_row`
-
-         """
-        return self._headers_cache.headers
-
-    @property
-    def header_indexes(self):
-        """Dictionary of header name: column index for this table's rows
-
-        Derived from :py:attr:`headers`
-
-        """
-        return self._headers_cache.indexes
-
-    def locate(self):
-        return sel.move_to_element(self._loc)
-
-    @property
-    def _root_loc(self):
-        return self.locate()
-
-    def rows(self):
-        """A generator method holding the Row objects
-
-        This generator yields Row objects starting at the first data row.
-
-        Yields:
-            :py:class:`Table.Row` object corresponding to the next row in the table.
-        """
-        index = self.body_offset
-        row_elements = sel.elements('tr', root=self.body)
-        for row_element in row_elements[index:]:
-            yield self.create_row_from_element(row_element)
-
-    def find_row(self, header, value):
-        """
-        Finds a row in the Table by iterating through each visible item.
-
-        Args:
-            header: A string or int, describing which column to inspect.
-            value: The value to be compared when trying to identify the correct row
-                to return.
-
-        Returns:
-            :py:class:`Table.Row` containing the requested cell, else ``None``.
-
-        """
-        return self.find_row_by_cells({header: value})
-
-    def find_cell(self, header, value):
-        """
-        Finds an item in the Table by iterating through each visible item,
-        this work used to be done by the :py:meth::`click_cell` method but
-        has not been abstracted out to be called separately.
-
-        Args:
-            header: A string or int, describing which column to inspect.
-            value: The value to be compared when trying to identify the correct cell
-                to click.
-
-        Returns: WebElement of the element if item was found, else ``None``.
-
-        """
-        matching_cell_rows = self.find_rows_by_cells({header: value})
-        try:
-            if isinstance(header, basestring):
-                return getattr(matching_cell_rows[0], header)
-            else:
-                return matching_cell_rows[0][header]
-        except IndexError:
-            return None
-
-    def find_rows_by_cells(self, cells, partial_check=False):
-        """A fast row finder, based on cell content.
-
-        Args:
-            cells: A dict of ``header: value`` pairs or a sequence of
-                nested ``(header, value)`` pairs.
-
-        Returns: A list of containing :py:class:`Table.Row` objects whose contents
-            match all of the header: value pairs in ``cells``
-
-        """
-        # accept dicts or supertuples
-        cells = dict(cells)
-        cell_text_loc = (
-            './/td/descendant-or-self::*[contains(normalize-space(text()), "%s")]/ancestor::tr[1]')
-        matching_rows_list = list()
-        for value in cells.values():
-            # Get all td elements that contain the value text
-            matching_elements = sel.elements(cell_text_loc % value,
-                root=sel.move_to_element(self._root_loc))
-            if matching_elements:
-                matching_rows_list.append(set(matching_elements))
-
-        # Now, find the common row elements that matched all the input cells
-        # (though not yet matching values to headers)
-        if not matching_rows_list:
-            # If none matched, short out
-            return []
-
-        rows_elements = list(reduce(lambda set1, set2: set1 & set2, matching_rows_list))
-
-        # Convert them to rows
-        # This is slow, which is why we do it after reducing the row element pile,
-        # and not when building matching_rows_list, but it makes comparing header
-        # names and expected values easy
-        rows = [self.create_row_from_element(element) for element in rows_elements]
-
-        # Only include rows where the expected values are in the right columns
-        matching_rows = list()
-        if partial_check:
-            matching_row_filter = lambda heading, value: value in row[heading].text
-        else:
-            matching_row_filter = lambda heading, value: row[heading].text == value
-        for row in rows:
-            if all(matching_row_filter(*cell) for cell in cells.items()):
-                matching_rows.append(row)
-
-        return matching_rows
-
-    def find_row_by_cells(self, cells, partial_check=False):
-        """Find the first row containing cells
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-
-        Returns: The first matching row found, or None if no matching row was found
-
-        """
-        try:
-            rows = self.find_rows_by_cells(cells, partial_check=partial_check)
-            return rows[0]
-        except IndexError:
-            return None
-
-    def click_rows_by_cells(self, cells, click_column=None, partial_check=False):
-        """Click the cell at ``click_column`` in the rows matched by ``cells``
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-            click_column: Which column in the row to click, defaults to None,
-                which will attempt to click the row element
-
-        Note:
-            The value of click_column can be a string or an int, and will be passed directly to
-            the item accessor (``__getitem__``) for :py:class:`Table.Row`
-
-        """
-        rows = self.find_rows_by_cells(cells, partial_check=partial_check)
-        if click_column is None:
-            map(sel.click, rows)
-        else:
-            map(sel.click, [row[click_column] for row in rows])
-
-    def click_row_by_cells(self, cells, click_column=None, partial_check=False):
-        """Click the cell at ``click_column`` in the first row matched by ``cells``
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-            click_column: See :py:meth:`Table.click_rows_by_cells`
-
-        """
-        row = self.find_row_by_cells(cells, partial_check=partial_check)
-        if click_column is None:
-            sel.click(row)
-        else:
-            sel.click(row[click_column])
-
-    def create_row_from_element(self, row_element):
-        """Given a row element in this table, create a :py:class:`Table.Row`
-
-        Args:
-            row_element: A table row (``<tr>``) WebElement representing a row in this table.
-
-        Returns: A :py:class:`Table.Row` for ``row_element``
-
-        """
-        return Table.Row(row_element, self)
-
-    def click_cells(self, cell_map):
-        """Submits multiple cells to be clicked on
-
-        Args:
-            cell_map: A mapping of header names and values, representing cells to click.
-                As an example, ``{'name': ['wing', 'nut']}, {'age': ['12']}`` would click on
-                the cells which had ``wing`` and ``nut`` in the name column and ``12`` in
-                the age column. The yaml example for this would be as follows::
-
-                    list_items:
-                        name:
-                            - wing
-                            - nut
-                        age:
-                            - 12
-
-        Raises:
-            NotAllItemsClicked: If some cells were unable to be found.
-
-        """
-        failed_clicks = []
-        for header, values in cell_map.items():
-            if isinstance(values, basestring):
-                values = [values]
-            for value in values:
-                res = self.click_cell(header, value)
-                if not res:
-                    failed_clicks.append("%s:%s" % (header, value))
-        if failed_clicks:
-            raise exceptions.NotAllItemsClicked(failed_clicks)
-
-    def click_cell(self, header, value):
-        """Clicks on a cell defined in the row.
-
-        Uses the header identifier and a value to determine which cell to click on.
-
-        Args:
-            header: A string or int, describing which column to inspect.
-            value: The value to be compared when trying to identify the correct cell
-                to click the cell in.
-
-        Returns: ``True`` if item was found and clicked, else ``False``.
-
-        """
-        cell = self.find_cell(header, value)
-        if cell:
-            sel.click(cell)
-            return True
-        else:
-            return False
-
-    class Row(Pretty):
-        """An object representing a row in a Table.
-
-        The Row object returns a dymanically addressable attribute space so that
-        the tables headers are automatically generated.
-
-        Args:
-            row_element: A table row ``WebElement``
-            parent_table: :py:class:`Table` containing ``row_element``
-
-        Notes:
-            Attributes are dynamically generated. The index/key accessor is more flexible
-            than the attr accessor, as it can operate on int indices and header names.
-
-        """
-        pretty_attrs = ['row_element', 'table']
-
-        def __init__(self, row_element, parent_table):
-            self.table = parent_table
-            self.row_element = row_element
-
-        @property
-        def columns(self):
-            """A list of WebElements corresponding to the ``<td>`` elements in this row"""
-            return sel.elements('td', root=self.row_element)
-
-        def __getattr__(self, name):
-            """
-            Returns Row element by header name
-            """
-            return self.columns[self.table.header_indexes[name.lower()]]
-
-        def __getitem__(self, index):
-            """
-            Returns Row element by header index or name
-            """
-            try:
-                return self.columns[index]
-            except TypeError:
-                # Index isn't an int, assume it's a string
-                return getattr(self, _convert_header(index))
-            # Let IndexError raise
-
-        def __str__(self):
-            return ", ".join(["'%s'" % el.text for el in self.columns])
-
-        def __eq__(self, other):
-            if isinstance(other, type(self)):
-                # Selenium elements support equality checks, so we can, too.
-                return self.row_element == other.row_element
-            else:
-                return id(self) == id(other)
-
-        def locate(self):
-            # table.create_row_from_element(row_instance) might actually work...
-            return sel.move_to_element(self.row_element)
-
-
-class SplitTable(Table):
-    """:py:class:`Table` that supports the header and body rows being in separate tables
-
-    Args:
-        header_data: A tuple, containing an element locator and an offset value.
-            These point to the container of the header row. The offset is used in case
-            there is a padding row above the header, or in the case that the header
-            and the body are contained inside the same table element.
-        body_data: A tuple, containing an element locator and an offset value.
-            These point to the container of the body rows. The offset is used in case
-            there is a padding row above the body rows, or in the case that the header
-            and the body are contained inside the same table element.
-
-    Usage:
-
-        table = SplitTable(header_data=('//div[@id="header_table"]//table/tbody', 0),
-            body_data=('//div[@id="body_table"]//table/tbody', 1))
-
-    The HTML code for a split table looks something like this::
-
-        <div id="prov_pxe_img_div">
-          <table id="header_table">
-              <tbody>
-                  <tr>
-                      <td>Name</td>
-                      <td>Animal</td>
-                      <td>Size</td>
-                  </tr>
-              </tbody>
-          </table>
-          <table id="body_table">
-              <tbody>
-                  <tr>
-                      <td>Useless</td>
-                      <td>Padding</td>
-                      <td>Row</td>
-                  </tr>
-                  <tr>
-                      <td>John</td>
-                      <td>Monkey</td>
-                      <td>Small</td>
-                  </tr>
-                  <tr>
-                      <td>Mike</td>
-                      <td>Tiger</td>
-                      <td>Large</td>
-                  </tr>
-              </tbody>
-          </table>
-        </div>
-
-    Note the use of the offset to skip the "Useless Padding Row" in ``body_data``. Most split
-    tables require an offset for both the heading and body rows.
-
-    """
-    def __init__(self, header_data, body_data):
-
-        self._header_loc, header_offset = header_data
-        self._body_loc, body_offset = body_data
-        self.header_offset = int(header_offset)
-        self.body_offset = int(body_offset)
-
-    @property
-    def _root_loc(self):
-        return self._body_loc
-
-    @property
-    def header_row(self):
-        """Property representing the ``<tr>`` element that contains header cells"""
-        # thead/tr containing header data
-        # xpath is 1-indexed, so we need to add 1 to the offset to get the correct row
-        return sel.element('tr[%d]' % (self.header_offset + 1), root=sel.element(self._header_loc))
-
-    @property
-    def body(self):
-        """Property representing the element that contains body rows"""
-        # tbody containing body rows
-        return sel.element(self._body_loc)
-
-    def locate(self):
-        # Use the header locator as the overall table locator
-        return sel.move_to_element(self._header_loc)
-
-
-class SortTable(Table):
-    """This table is the same as :py:class:`Table`, but with added sorting functionality."""
-    @property
-    def _sort_by_cell(self):
-        try:
-            return sel.element("./th/a/img[contains(@src, 'sort')]/..", root=self.header_row)
-        except NoSuchElementException:
-            return None
-
-    @property
-    def sorted_by(self):
-        """Return column name what is used for sorting now.
-        """
-        cell = self._sort_by_cell
-        if cell is None:
-            return None
-        return sel.text("./a", root=cell).encode("utf-8")
-
-    @property
-    def sort_order(self):
-        """Return order.
-
-        Returns: 'ascending' or 'descending'
-        """
-        cell = self._sort_by_cell
-        if cell is None:
-            return None
-
-        def _downstream():
-            src = sel.get_attribute(sel.element("./img", root=cell), "src")
-            if "sort_up" in src:
-                return "ascending"
-            elif "sort_down" in src:
-                return "descending"
-            else:
-                return None
-
-        def _upstream():
-            cls = sel.get_attribute(cell, "class")
-            if "sorting_asc" in cls:
-                return "ascending"
-            elif "sorting_desc" in cls:
-                return "descending"
-            else:
-                return None
-
-        return version.pick({
-            "default": _downstream,
-            "5.3.0.0": _upstream
-        })()
-
-    def click_header_cell(self, text):
-        """Clicks on the header to change sorting conditions.
-
-        Args:
-            text: Header cell text.
-        """
-        sel.click(sel.element("./th/a[normalize-space(.)='{}']".format(text), root=self.header_row))
-
-    def sort_by(self, header, order):
-        """Sorts the table by given conditions
-
-        Args:
-            header: Text of the header cell to use for sorting.
-            order: ascending or descending
-        """
-        order = order.lower().strip()
-        if header != self.sorted_by:
-            # Change column to order by
-            self.click_header_cell(header)
-            assert self.sorted_by == header, "Detected malfunction in table ordering"
-        if order != self.sort_order:
-            # Change direction
-            self.click_header_cell(header)
-            assert self.sort_order == order, "Detected malfunction in table ordering"
-
-
-class CheckboxTable(Table):
-    """:py:class:`Table` with support for checkboxes
-
-    Args:
-        table_locator: See :py:class:`cfme.web_ui.Table`
-        header_checkbox_locator: Locator of header checkbox (default `None`)
-                                 Specify in case the header checkbox is not part of the header row
-        body_checkbox_locator: Locator for checkboxes in body rows
-        header_offset: See :py:class:`cfme.web_ui.Table`
-        body_offset: See :py:class:`cfme.web_ui.Table`
-    """
-    _checkbox_loc = ".//input[@type='checkbox']"
-
-    def __init__(self, table_locator, header_offset=0, body_offset=0,
-            header_checkbox_locator=None, body_checkbox_locator=None):
-        super(CheckboxTable, self).__init__(table_locator, header_offset, body_offset)
-        if body_checkbox_locator:
-            self._checkbox_loc = body_checkbox_locator
-        self._header_checkbox_loc = header_checkbox_locator
-
-    @property
-    def header_checkbox(self):
-        """Checkbox used to select/deselect all rows"""
-        if self._header_checkbox_loc is not None:
-            return sel.element(self._header_checkbox_loc)
-        else:
-            return sel.element(self._checkbox_loc, root=self.header_row)
-
-    def select_all(self):
-        """Select all rows using the header checkbox or one by one if not present"""
-        if self._header_checkbox_loc is None:
-            for row in self.rows():
-                self._set_row_checkbox(row, True)
-        else:
-            sel.uncheck(self.header_checkbox)
-            sel.check(self.header_checkbox)
-
-    def deselect_all(self):
-        """Deselect all rows using the header checkbox or one by one if not present"""
-        if self._header_checkbox_loc is None:
-            for row in self.rows():
-                self._set_row_checkbox(row, False)
-        else:
-            sel.check(self.header_checkbox)
-            sel.uncheck(self.header_checkbox)
-
-    def _set_row_checkbox(self, row, set_to=False):
-        row_checkbox = sel.element(self._checkbox_loc, root=row.locate())
-        sel.checkbox(row_checkbox, set_to)
-
-    def _set_row(self, header, value, set_to=False):
-        """ Internal method used to select/deselect a row by column header and cell value
-
-        Args:
-            header: See :py:meth:`Table.find_row`
-            value: See :py:meth:`Table.find_row`
-            set_to: Select if `True`, deselect if `False`
-        """
-        row = self.find_row(header, value)
-        if row:
-            self._set_row_checkbox(row, set_to)
-            return True
-        else:
-            return False
-
-    def select_rows_by_indexes(self, *indexes):
-        """Select rows specified by row indexes (starting with 0)
-        """
-        for i, row in enumerate(self.rows()):
-            if i in indexes:
-                self._set_row_checkbox(row, True)
-
-    def deselect_rows_by_indexes(self, *indexes):
-        """Deselect rows specified by row indexes (starting with 0)
-        """
-        for i, row in enumerate(self.rows()):
-            if i in indexes:
-                self._set_row_checkbox(row, False)
-
-    def select_row(self, header, value):
-        """Select a single row specified by column header and cell value
-
-        Args:
-            header: See :py:meth:`Table.find_row`
-            value: See :py:meth:`Table.find_row`
-
-        Returns: `True` if successful, `False` otherwise
-        """
-        return self._set_row(header, value, True)
-
-    def deselect_row(self, header, value):
-        """Deselect a single row specified by column header and cell value
-
-        Args:
-            header: See :py:meth:`Table.find_row`
-            value: See :py:meth:`Table.find_row`
-
-        Returns: `True` if successful, `False` otherwise
-        """
-        return self._set_row(header, value, False)
-
-    def _set_rows(self, cell_map, set_to=False):
-        """ Internal method used to select/deselect multiple rows
-
-        Args:
-            cell_map: See :py:meth:`Table.click_cells`
-            set_to: Select if `True`, deselect if `False`
-        """
-        failed_selects = []
-        for header, values in cell_map.items():
-            if isinstance(values, basestring):
-                values = [values]
-            for value in values:
-                res = self._set_row(header, value, set_to)
-                if not res:
-                    failed_selects.append("%s:%s" % (header, value))
-        if failed_selects:
-            raise exceptions.NotAllCheckboxesFound(failed_selects)
-
-    def select_rows(self, cell_map):
-        """Select multiple rows
-
-        Args:
-            cell_map: See :py:meth:`Table.click_cells`
-
-        Raises:
-            NotAllCheckboxesFound: If some cells were unable to be found
-        """
-        self._set_rows(cell_map, True)
-
-    def deselect_rows(self, cell_map):
-        """Deselect multiple rows
-
-        Args:
-            cell_map: See :py:meth:`Table.click_cells`
-
-        Raises:
-            NotAllCheckboxesFound: If some cells were unable to be found
-        """
-        self._set_rows(cell_map, False)
-
-    def _set_row_by_cells(self, cells, set_to=False):
-        row = self.find_row_by_cells(cells)
-        self._set_row_checkbox(row, set_to)
-
-    def select_row_by_cells(self, cells):
-        """Select the first row matched by ``cells``
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-
-        """
-        self._set_row_by_cells(cells, True)
-
-    def deselect_row_by_cells(self, cells):
-        """Deselect the first row matched by ``cells``
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-
-        """
-        self._set_row_by_cells(cells, False)
-
-    def _set_rows_by_cells(self, cells, set_to=False):
-        rows = self.find_rows_by_cells(cells)
-        for row in rows:
-            self._set_row_checkbox(row, set_to)
-
-    def select_rows_by_cells(self, cells):
-        """Select the rows matched by ``cells``
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-        """
-        self._set_rows_by_cells(cells, True)
-
-    def deselect_rows_by_cells(self, cells):
-        """Deselect the rows matched by ``cells``
-
-        Args:
-            cells: See :py:meth:`Table.find_rows_by_cells`
-        """
-        self._set_rows_by_cells(cells, False)
-
-
-class SplitCheckboxTable(SplitTable, CheckboxTable):
-    """:py:class:`SplitTable` with support for checkboxes
-
-    Args:
-        header_data: See :py:class:`cfme.web_ui.SplitTable`
-        body_data: See :py:class:`cfme.web_ui.SplitTable`
-        header_checkbox_locator: See :py:class:`cfme.web_ui.CheckboxTable`
-        body_checkbox_locator: See :py:class:`cfme.web_ui.CheckboxTable`
-        header_offset: See :py:class:`cfme.web_ui.Table`
-        body_offset: See :py:class:`cfme.web_ui.Table`
-    """
-    _checkbox_loc = './/img[contains(@src, "item_chk")]'
-
-    def __init__(self, header_data, body_data,
-            header_checkbox_locator=None, body_checkbox_locator=None):
-        # To limit multiple inheritance surprises, explicitly call out to SplitTable's __init__
-        SplitTable.__init__(self, header_data, body_data)
-
-        # ...then set up CheckboxTable's locators here
-        self._header_checkbox_loc = header_checkbox_locator
-        if body_checkbox_locator:
-            self._checkbox_loc = body_checkbox_locator
-
-
-class PagedTable(Table):
-    """:py:class:`Table` with support for paginator
-
-    Args:
-        table_locator: See :py:class:`cfme.web_ui.Table`
-        header_checkbox_locator: Locator of header checkbox (default `None`)
-                                 Specify in case the header checkbox is not part of the header row
-        body_checkbox_locator: Locator for checkboxes in body rows
-        header_offset: See :py:class:`cfme.web_ui.Table`
-        body_offset: See :py:class:`cfme.web_ui.Table`
-    """
-    def find_row_on_all_pages(self, header, value):
-        from cfme.web_ui import paginator
-        for page in paginator.pages():
-            sel.wait_for_element(self)
-            row = self.find_row(header, value)
-            if row is not None:
-                return row
-
-
-class SplitPagedTable(SplitTable, PagedTable):
-    """:py:class:`SplitTable` with support for paginator
-
-    Args:
-        header_data: See :py:class:`cfme.web_ui.SplitTable`
-        body_data: See :py:class:`cfme.web_ui.SplitTable`
-        header_offset: See :py:class:`cfme.web_ui.Table`
-        body_offset: See :py:class:`cfme.web_ui.Table`
-    """
-    def __init__(self, header_data, body_data):
-        # To limit multiple inheritance surprises, explicitly call out to SplitTable's __init__
-        SplitTable.__init__(self, header_data, body_data)
-
-
-def table_in_object(table_title):
-    """If you want to point to tables inside object view, this is what you want to use.
-
-    Works both on down- and upstream.
-
-    Args:
-        table_title: Text in `p` element preceeding the table
-    Returns: XPath locator for the desired table.
-    """
-    return ("//table[(preceding-sibling::p[1] | preceding-sibling::h3[1])[normalize-space(.)={}]]"
-        .format(quoteattr(table_title)))
-
-
 @multimethod(lambda loc, value: (sel.tag(loc), sel.get_attribute(loc, 'type')))
 def fill_tag(loc, value):
     """ Return a tuple of function to do the filling, and a value to log."""
